# docs/swagger.yaml
basePath: /
definitions:
  http_user_requests.LoginRequest:
    properties:
      email:
        type: string
    type: object
  http_user_requests.RefreshRequest:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  http_user_requests.RegisterRequest:
    properties:
      email:
        type: string
    type: object
  http_user_responses.LoginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  http_user_responses.RefreshResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  http_user_responses.RegisterResponse:
    properties:
      user:
        $ref: '#/definitions/http_user_structs.User'
    type: object
  http_user_structs.User:
    properties:
      email:
        type: string
      id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  title: Authentication Service API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Logs in a user using their user_id query parameter and returns
        an access token and refresh token.
      operationId: login-user
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      - description: User login details
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/http_user_requests.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful login
          schema:
            $ref: '#/definitions/http_user_responses.LoginResponse'
        "400":
          description: Bad Request - missing or invalid user_id
          schema:
            type: string
        "404":
          description: Not Found - user not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refreshes the access token using a valid refresh token.
      operationId: refresh-token
      parameters:
      - description: Refresh token details
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/http_user_requests.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Tokens refreshed successfully
          schema:
            $ref: '#/definitions/http_user_responses.RefreshResponse'
        "400":
          description: Bad Request - invalid input
          schema:
            type: string
        "401":
          description: Unauthorized - invalid refresh token
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
  /auth/register:
    post:
      consumes:
      - application/json
      description: Registers a new user by providing an email address and other necessary
        information.
      operationId: register-user
      parameters:
      - description: User registration details
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/http_user_requests.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully registered new user
          schema:
            $ref: '#/definitions/http_user_responses.RegisterResponse'
        "400":
          description: Bad Request - invalid input
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
swagger: "2.0"